+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Updated the chalk dependency (no change in behavior)
 - (Internal) Updated the dev dependencies, involving lot of small changes adding parentheses to lambda parameters.
+- Fix --allowUnusedTypes not applying to type re-exports (Issue #180)
+- Fix --allowUnusedEnums not applying to enum re-exports
 
 ## [7.0.0] - 11 Nov 2020